backend/ai_service/face_service/requirements.txt
flask
deepface
mysql-connector-python
numpy
flask-cors
tf-keras
mediapipe
opencv-python
